Move Visitors container
