+++ front/app/containers/Admin/dashboard/visitors/VisitorsTypeCard/index.tsx
