+++ front/app/containers/Admin/dashboard/visitors/VisitorsTypeCard/useVisitorTypes/parse.ts
